Add many of the S loop rules
Also managed to name some of the implicit groupings. And the phonetic
section definitely explains the rules and even character selection a
good bit more, as well as providing a vocabulary for the different
shapes. Though I'm sticking with "curved foot" over "up-hook" since
the up-hook is on the down side (exits up), which could be really
confused. Head/foot is not so confusable.