Fill in "Marks" section
Attempted to match the style of the rest of the document, but it is certainly in need of review.  Particularly the John Thomas line.  In the scan of the manual there appears to be a line break continuation character in the shorthand, and I wasn't sure if that was required.